claptrap is now expecting a path to a .so file when executed.
Remove the logic of a yaml file describing  what action should be performed on a specific caught event.
To use claptrap more freely, I rather expose a simple API and let the users implement their own logic.
Claptrap is now  relying on go plugin and expect to receive as argument the path towards a shared object
it will lookup for a symbol name 'Handle' and ensure that the exported function has the  following signature:

	func(string, string, string)

This 'Handle' function embedding user's logic is executed in a separate goroutine for each event detected.